backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary>=2.9.10
djangorestframework-simplejwt>=5.4.0
django-cors-headers==4.3.0
Pillow>=10.4.0
python-decouple==3.8
django-filter>=24.0
drf-yasg==1.21.7
setuptools>=65.0.0

# Production сервер
gunicorn==21.2.0
whitenoise==6.6.0

